Unify campaign testing fakery.
