+++ enterprise/internal/campaigns/webhooks_test.go
@@ -9,7 +9,6 @@ import (
 	"encoding/hex"
 	"encoding/json"
 	"flag"
-	"io"
 	"io/ioutil"
 	"net/http"
 	"net/http/httptest"
@@ -23,13 +22,12 @@ import (
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/go-cmp/cmp/cmpopts"
 	"github.com/sourcegraph/sourcegraph/cmd/repo-updater/repos"
+	ct "github.com/sourcegraph/sourcegraph/enterprise/internal/campaigns/testing"
 	"github.com/sourcegraph/sourcegraph/internal/campaigns"
 	"github.com/sourcegraph/sourcegraph/internal/extsvc"
 	"github.com/sourcegraph/sourcegraph/internal/httptestutil"
 	"github.com/sourcegraph/sourcegraph/internal/rcache"
-	"github.com/sourcegraph/sourcegraph/internal/repoupdater"
 	"github.com/sourcegraph/sourcegraph/internal/repoupdater/protocol"
-	"github.com/sourcegraph/sourcegraph/internal/vcs/git"
 	"github.com/sourcegraph/sourcegraph/schema"
 )
 
@@ -115,21 +113,11 @@ func testGitHubWebhook(db *sql.DB, userID int32) func(*testing.T) {
 		// Set up mocks to prevent the diffstat computation from trying to
 		// use a real gitserver, and so we can control what diff is used to
 		// create the diffstat.
-		repoupdater.MockRepoLookup = func(args protocol.RepoLookupArgs) (*protocol.RepoLookupResult, error) {
-			return &protocol.RepoLookupResult{
-				Repo: &protocol.RepoInfo{
-					Name: "repo",
-					VCS:  protocol.VCSInfo{URL: "https://example.com/repo/"},
-				},
-			}, nil
-		}
-		git.Mocks.ExecReader = func(args []string) (io.ReadCloser, error) {
-			return ioutil.NopCloser(strings.NewReader("")), nil
-		}
-		defer func() {
-			repoupdater.MockRepoLookup = nil
-			git.ResetMocks()
-		}()
+		state := ct.MockGitHubChangesetSync(&protocol.RepoInfo{
+			Name: "repo",
+			VCS:  protocol.VCSInfo{URL: "https://example.com/repo/"},
+		})
+		defer state.Unmock()
 
 		err = SyncChangesets(ctx, repoStore, store, cf, changesets...)
 		if err != nil {
